finished student features

+++ src/main/java/home/jakartasubmit/SubmissionServlet.java
@@ -15,6 +15,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
+import static home.jakartasubmit.models.User.Role.ADMIN;
+
 @WebServlet(name = "SubmissionServlet", value = "/submission-servlet")
 @MultipartConfig
 public class SubmissionServlet extends HttpServlet {
@@ -79,22 +81,26 @@ public class SubmissionServlet extends HttpServlet {
         List<Submission> submissions = new ArrayList<>();
         List<Task> tasks = new ArrayList<>();
         List<User> students = new ArrayList<>();
+        String page = "error.jsp";
 
         switch (role) {
             case ADMIN:
                 submissions = submissionService.getAllSubmissions();
                 tasks = taskService.getAllTasks();
+                page = "/admin/Submissions.jsp";
                 break;
             case INSTRUCTOR:
                 User instructor = userService.getUserByEmail(currentUserDTO.getEmail());
                 submissions = submissionService.getAllSubmissions();
                 tasks = taskService.getTasksByInstructor(instructor);
+                page = "/instructor/Submissions.jsp";
                 break;
             case STUDENT:
                 User student = userService.getUserByEmail(currentUserDTO.getEmail());
                 submissions = submissionService.getSubmissionsByStudent(student);
                 students = userService.getUsersByRole(role);
                 tasks = taskService.getAllTasks();
+                page = "/student/Submissions.jsp";
                 break;
             default:
                 response.sendRedirect("error.jsp"); // Handle unexpected role case
@@ -110,11 +116,11 @@ public class SubmissionServlet extends HttpServlet {
 
             request.setAttribute("message", "Submissions got successfully");
             request.setAttribute("messageType", "success");
-            request.getRequestDispatcher("public/Submissions.jsp").forward(request, response);
+            request.getRequestDispatcher(page).forward(request, response);
         } else {
             request.setAttribute("message", "Error getting submissions");
             request.setAttribute("messageType", "danger");
-            request.getRequestDispatcher("submission.jsp").forward(request, response);
+            request.getRequestDispatcher(page).forward(request, response);
         }
     }
 
@@ -141,6 +147,14 @@ public class SubmissionServlet extends HttpServlet {
             currentUserDTO = (UserDTO) sessionobj.getAttribute("currentUser");
         }
 
+        String page = "";
+        page = switch (currentUserDTO.getRole()) {
+            case ADMIN -> "/admin/Submissions.jsp";
+            case INSTRUCTOR -> "/instructor/Submissions.jsp";
+            case STUDENT -> "/student/Submissions.jsp";
+            default -> page = "/error.jsp";
+        };
+
         UUID taskId = UUID.fromString(request.getParameter("taskId"));
         String filePath = "";
         Part filePart = request.getPart("submissionFile"); // Ensure the form has enctype="multipart/form-data"
@@ -148,7 +162,7 @@ public class SubmissionServlet extends HttpServlet {
         try{
             // Ensure file is present
             if (filePart == null || filePart.getSize() == 0) {
-                response.getWriter().write("No file uploaded. <a href=\"" + request.getContextPath() + "/public/Submissions.jsp\">Return back</a>");
+                response.getWriter().write("No file uploaded. <a href="+ request.getContextPath() + page +">Return back</a>");
                 return;
             }
 
@@ -157,7 +171,6 @@ public class SubmissionServlet extends HttpServlet {
 
             User student = userService.getUserByEmail(currentUserDTO.getEmail());
             Task task = taskService.getTaskById(taskId);
-
 
             // Check if both student and task are found
             if (student != null && task != null) {
@@ -168,12 +181,12 @@ public class SubmissionServlet extends HttpServlet {
                 // Register the submission via the service
                 if (submissionService.registerSubmission(submission)) {
                     // let's save the
-                    response.getWriter().write("Submission created successfully!, <a href=\"" + request.getContextPath() + "/submission\">Return back</a>");
+                    response.getWriter().write("Submission created successfully!, <a href=" + request.getContextPath()+ page + ">Return back</a>");
                 } else {
-                    response.getWriter().write("Submission creation failed. <a href=\"" + request.getContextPath() + "/submission\">Return back</a>");
+                    response.getWriter().write("Submission creation failed. <a href=" + request.getContextPath()+page + ">Return back</a>");
                 }
             } else {
-                response.getWriter().write("Invalid student or task. <a href=\"" + request.getContextPath() + "/submission\">Return back</a>");
+                response.getWriter().write("Invalid student or task. <a href=" + request.getContextPath()+page + ">Return back</a>");
             }
         } catch (Exception e){
             response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);  // Set status code to 500
@@ -196,6 +209,14 @@ public class SubmissionServlet extends HttpServlet {
             response.getWriter().write("Unauthorized access. <a href=\"login.jsp\">Return to Login</a>");
             return;
         }
+
+        String page = "";
+        page = switch (currentUserDTO.getRole()) {
+            case ADMIN -> "/admin/Submissions.jsp";
+            case INSTRUCTOR -> "/instructor/Submissions.jsp";
+            case STUDENT -> "/student/Submissions.jsp";
+            default -> page = "/error.jsp";
+        };
 
         try {
             UUID submissionId = UUID.fromString(request.getParameter("submissionId"));
@@ -223,17 +244,37 @@ public class SubmissionServlet extends HttpServlet {
         }
 
         // Return link
-        response.getWriter().write("<br/><a href=\"/Jakarta-Submit-1.0-SNAPSHOT/submission\">Return back</a>");
+        response.getWriter().write("<br/><a href="+ request.getContextPath()+page +">Return back</a>");
     }
 
     private void handleSubmissionDeletion(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         response.setContentType("text/html; charset=UTF-8");
 
         UUID submissionId = UUID.fromString(request.getParameter("submissionId"));
+
+        HttpSession sessionobj = request.getSession(false);
+        UserDTO currentUserDTO = null;
+
+        if (sessionobj != null && sessionobj.getAttribute("isLoggedIn") != null) {
+            currentUserDTO = (UserDTO) sessionobj.getAttribute("currentUser");
+        }
+
+        if (currentUserDTO == null) {
+            response.getWriter().write("Unauthorized access. <a href=\"login.jsp\">Return to Login</a>");
+            return;
+        }
+
+        String page = "";
+        page = switch (currentUserDTO.getRole()) {
+            case ADMIN -> "/admin/Submissions.jsp";
+            case INSTRUCTOR -> "/instructor/Submissions.jsp";
+            case STUDENT -> "/student/Submissions.jsp";
+            default -> page = "/error.jsp";
+        };
         if (submissionService.deleteSubmission(submissionId)) {
-            response.getWriter().write("Submission deleted successfully. <a href=\"" + request.getContextPath() + "/submission\">Return back</a>");
+            response.getWriter().write("Submission deleted successfully. <a href=" + request.getContextPath() + ">Return back</a>");
         } else {
-            response.getWriter().write("Error deleting submission. <a href=\"" + request.getContextPath() + "/submission\">Return back</a>");
+            response.getWriter().write("Error deleting submission. <a href=" + request.getContextPath()+page + ">Return back</a>");
         }
     }
 }
